added Command ServerId, ClientId functioalities

+++ helper/helper.go
@@ -13,8 +13,8 @@ func TimeNow() int {
 	return int(time.Now().Unix())
 }
 
-func TimeNowNano() int {
-    return int(time.Now().UnixNano())
+func TimeNowNano() int64 {
+    return time.Now().UnixNano()
 }
 func TimeNowMs() int {
 	return int(time.Now().UnixNano() / 1000000)
